# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API epesantren pegawai",
        "description": "Dokumentasi API epesantren pegawai",
        "contact": {
            "email": "fp3175723@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/auth/teaching-history": {
            "get": {
                "tags": [
                    "Profil_Riwayat_hidup-Teaching"
                ],
                "summary": "Get data teaching",
                "description": "Mengambil semua data teaching berdasarkan user login",
                "operationId": "5ffdbfd4324d4c3c84d1541d4ab0203b",
                "responses": {
                    "200": {
                        "description": "Data teaching berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "teaching_id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "teaching_start": {
                                                        "type": "string",
                                                        "example": "2000"
                                                    },
                                                    "teaching_end": {
                                                        "type": "string",
                                                        "example": "2024"
                                                    },
                                                    "teaching_lesson": {
                                                        "type": "string",
                                                        "example": "IPA"
                                                    },
                                                    "teaching_desc": {
                                                        "type": "string",
                                                        "example": "ilmu pengetahuan alam"
                                                    },
                                                    "teaching_employee_id": {
                                                        "type": "integer",
                                                        "example": "118"
                                                    },
                                                    "sekolah_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "null",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profil_Riwayat_hidup-Teaching"
                ],
                "summary": "Membuat data employee-fam",
                "description": "Mengirim data teaching history",
                "operationId": "94628c2345b667bb9d0dc72948046029",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "teaching_start",
                                    "teaching_end",
                                    "teaching_lesson",
                                    "teaching_desc"
                                ],
                                "properties": {
                                    "teaching_start": {
                                        "type": "string",
                                        "example": "2010-09-30"
                                    },
                                    "teaching_end": {
                                        "type": "string",
                                        "example": "2010-09-30"
                                    },
                                    "teaching_lesson": {
                                        "type": "string",
                                        "example": "IPA"
                                    },
                                    "teaching_desc": {
                                        "type": "string",
                                        "example": "pertukangan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data teaching history berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validasi data  gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/teaching-history/{teaching_id}": {
            "delete": {
                "tags": [
                    "Profil_Riwayat_hidup-Teaching"
                ],
                "summary": "Delete teaching history",
                "description": "Menghapus data teaching history berdasarkan ID",
                "operationId": "86e906d86a9d49630361ced7fa3a9ef2",
                "parameters": [
                    {
                        "name": "teaching_id",
                        "in": "path",
                        "description": "ID family untuk menghapus data ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "data berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Data tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "data tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Profil_Riwayat_hidup-Teaching"
                ],
                "summary": "update data teaching history",
                "description": "Update data teaching history",
                "operationId": "cc3e3a0454d9c491c3160c58fd5b5143",
                "parameters": [
                    {
                        "name": "teaching_id",
                        "in": "path",
                        "description": "ID untuk memperbarui data",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "teaching_start",
                                    "teaching_end",
                                    "teaching_lesson",
                                    "teaching_desc"
                                ],
                                "properties": {
                                    "teaching_start": {
                                        "type": "string",
                                        "example": "2010-09-30"
                                    },
                                    "teaching_end": {
                                        "type": "string",
                                        "example": "2010-09-30"
                                    },
                                    "teaching_lesson": {
                                        "type": "string",
                                        "example": "IPA"
                                    },
                                    "teaching_desc": {
                                        "type": "string",
                                        "example": "pertukangan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data teaching history berhasil diupate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validasi data gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}